Move `updateQuestion` action to its own file
