+++ frontend/src/metabase/query_builder/actions/core/core.js
@@ -1,5 +1,4 @@
 import _ from "underscore";
-import { assocIn } from "icepick";
 import { createAction } from "redux-actions";
 
 import * as MetabaseAnalytics from "metabase/lib/analytics";
@@ -20,54 +19,27 @@ import { setRequestUnloaded } from "metabase/redux/requests";
 import Questions from "metabase/entities/questions";
 import { fetchAlertsForQuestion } from "metabase/alert/alert";
 
-import StructuredQuery from "metabase-lib/lib/queries/StructuredQuery";
-
-import {
-  getTemplateTagsForParameters,
-  getTemplateTagParameters,
-} from "metabase/parameters/utils/cards";
-
 import { trackNewQuestionSaved } from "../../analytics";
 import {
   getCard,
-  getFirstQueryResult,
-  getIsEditing,
-  getIsShowingTemplateTagsEditor,
   getOriginalQuestion,
-  getQueryBuilderMode,
   getQuestion,
-  getRawSeries,
   getResultsMetadata,
   getTransformedSeries,
   isBasedOnExistingQuestion,
 } from "../../selectors";
-import { getNextTemplateTagVisibilityState } from "../../utils";
 
 import { updateUrl } from "../navigation";
-import { setIsShowingTemplateTagsEditor } from "../native";
 import { zoomInRow } from "../object-detail";
 import { clearQueryResult, runQuestionQuery } from "../querying";
-import { onCloseSidebars, setQueryBuilderMode } from "../ui";
+import { onCloseSidebars } from "../ui";
 
-import { loadMetadataForCard } from "./metadata";
-import { getQuestionWithDefaultVisualizationSettings } from "./utils";
-
+import { updateQuestion } from "./updateQuestion";
 import { loadMetadataForCard } from "./metadata";
 import { getQuestionWithDefaultVisualizationSettings } from "./utils";
 
 export const RESET_QB = "metabase/qb/RESET_QB";
 export const resetQB = createAction(RESET_QB);
-
-function hasNewColumns(question, queryResult) {
-  // NOTE: this assume column names will change
-  // technically this is wrong because you could add and remove two columns with the same name
-  const query = question.query();
-  const previousColumns =
-    (queryResult && queryResult.data.cols.map(col => col.name)) || [];
-  const nextColumns =
-    query instanceof StructuredQuery ? query.columnNames() : [];
-  return _.difference(nextColumns, previousColumns).length > 0;
-}
 
 // refreshes the card without triggering a run of the card's query
 export const SOFT_RELOAD_CARD = "metabase/qb/SOFT_RELOAD_CARD";
@@ -182,166 +154,6 @@ export const navigateToNewCardInsideQB = createThunkAction(
     };
   },
 );
-
-/**
- * Replaces the currently active question with the given Question object.
- * Also shows/hides the template tag editor if the number of template tags has changed.
- */
-export const UPDATE_QUESTION = "metabase/qb/UPDATE_QUESTION";
-export const updateQuestion = (
-  newQuestion,
-  {
-    shouldStartAdHocQuestion = true,
-    run = false,
-    shouldUpdateUrl = false,
-  } = {},
-) => {
-  return async (dispatch, getState) => {
-    const oldQuestion = getQuestion(getState());
-    const mode = getQueryBuilderMode(getState());
-
-    // TODO Atte Keinänen 6/2/2017 Ways to have this happen automatically when modifying a question?
-    // Maybe the Question class or a QB-specific question wrapper class should know whether it's being edited or not?
-    if (
-      shouldStartAdHocQuestion &&
-      newQuestion.query().isEditable() &&
-      !getIsEditing(getState()) &&
-      newQuestion.isSaved() &&
-      mode !== "dataset"
-    ) {
-      newQuestion = newQuestion.withoutNameAndId();
-
-      // When the dataset query changes, we should loose the dataset flag,
-      // to start building a new ad-hoc question based on a dataset
-      if (newQuestion.isDataset()) {
-        newQuestion = newQuestion.setDataset(false);
-      }
-    }
-
-    const queryResult = getFirstQueryResult(getState());
-    newQuestion = newQuestion.syncColumnsAndSettings(oldQuestion, queryResult);
-
-    if (run === "auto") {
-      run = hasNewColumns(newQuestion, queryResult);
-    }
-
-    if (!newQuestion.canAutoRun()) {
-      run = false;
-    }
-
-    // <PIVOT LOGIC>
-    // We have special logic when going to, coming from, or updating a pivot table.
-    const isPivot = newQuestion.display() === "pivot";
-    const wasPivot = oldQuestion.display() === "pivot";
-    const queryHasBreakouts =
-      isPivot &&
-      newQuestion.isStructured() &&
-      newQuestion.query().breakouts().length > 0;
-
-    // we can only pivot queries with breakouts
-    if (isPivot && queryHasBreakouts) {
-      // compute the pivot setting now so we can query the appropriate data
-      const series = assocIn(
-        getRawSeries(getState()),
-        [0, "card"],
-        newQuestion.card(),
-      );
-      const key = "pivot_table.column_split";
-      const setting = getQuestionWithDefaultVisualizationSettings(
-        newQuestion,
-        series,
-      ).setting(key);
-      newQuestion = newQuestion.updateSettings({ [key]: setting });
-    }
-
-    if (
-      // switching to pivot
-      (isPivot && !wasPivot && queryHasBreakouts) ||
-      // switching away from pivot
-      (!isPivot && wasPivot) ||
-      // updating the pivot rows/cols
-      (isPivot &&
-        queryHasBreakouts &&
-        !_.isEqual(
-          newQuestion.setting("pivot_table.column_split"),
-          oldQuestion.setting("pivot_table.column_split"),
-        ))
-    ) {
-      run = true; // force a run when switching to/from pivot or updating it's setting
-    }
-    // </PIVOT LOGIC>
-
-    // Native query should never be in notebook mode (metabase#12651)
-    if (mode === "notebook" && newQuestion.isNative()) {
-      await dispatch(
-        setQueryBuilderMode("view", {
-          shouldUpdateUrl: false,
-        }),
-      );
-    }
-
-    const newDatasetQuery = newQuestion.query().datasetQuery();
-    // Sync card's parameters with the template tags;
-    if (newDatasetQuery.type === "native") {
-      const templateTags = getTemplateTagsForParameters(newQuestion.card());
-      const parameters = getTemplateTagParameters(templateTags);
-      newQuestion = newQuestion.setParameters(parameters);
-    }
-
-    // Replace the current question with a new one
-    await dispatch.action(UPDATE_QUESTION, { card: newQuestion.card() });
-
-    if (shouldUpdateUrl) {
-      dispatch(updateUrl(null, { dirty: true }));
-    }
-
-    // See if the template tags editor should be shown/hidden
-    const isTemplateTagEditorVisible = getIsShowingTemplateTagsEditor(
-      getState(),
-    );
-    const nextTagEditorVisibilityState = getNextTemplateTagVisibilityState({
-      oldQuestion,
-      newQuestion,
-      isTemplateTagEditorVisible,
-      queryBuilderMode: mode,
-    });
-    if (nextTagEditorVisibilityState !== "deferToCurrentState") {
-      dispatch(
-        setIsShowingTemplateTagsEditor(
-          nextTagEditorVisibilityState === "visible",
-        ),
-      );
-    }
-
-    try {
-      if (
-        !_.isEqual(
-          oldQuestion.query().dependentMetadata(),
-          newQuestion.query().dependentMetadata(),
-        )
-      ) {
-        await dispatch(loadMetadataForCard(newQuestion.card()));
-      }
-
-      // setDefaultQuery requires metadata be loaded, need getQuestion to use new metadata
-      const question = getQuestion(getState());
-      const questionWithDefaultQuery = question.setDefaultQuery();
-      if (!questionWithDefaultQuery.isEqual(question)) {
-        await dispatch.action(UPDATE_QUESTION, {
-          card: questionWithDefaultQuery.setDefaultDisplay().card(),
-        });
-      }
-    } catch (e) {
-      // this will fail if user doesn't have data permissions but thats ok
-      console.warn("Couldn't load metadata", e);
-    }
-
-    // run updated query
-    if (run) {
-      dispatch(runQuestionQuery());
-    }
-  };
-};
 
 // DEPRECATED, still used in a couple places
 export const setDatasetQuery =
